feat(Web-UI): Added cancel button to the MediaOverview screen when refreshing
